// src/app/modules/liquidaciones/models/resultado-hisopado.model.ts
export class ResultadoLaboratorio {
    liquidacionId: number;
    incidenteId: number;
    fechaHora: Date;
    fechaHoraString: string;
    resultado: number;
    motivoNoRealizado: number;
    laboratorioId: number;
    observaciones: string;
    adjuntoFile: string;
    adjuntoFileName: string;
    archivoOrden: string;
}